1)lsh functions implemented 2) minor changes of var types in myHashTable 3) added utils.cpp and utils.h for universal use of calculate_distance and  vector_multiply

+++ myHashTable.h
@@ -22,7 +22,7 @@ class myHashTable {
         vector<float> tParameters;
 
         //a hash table of sets
-        set< pair< vector<float>*,unsigned int > >* myHash;
+        set< pair< vector<float>*,int > >* myHash;
 
         //count of buckets
         unsigned int myHashSize;
@@ -32,18 +32,18 @@ class myHashTable {
         myHashTable(unsigned int k, unsigned int N, unsigned int w, unsigned int dimensions);
         ~myHashTable();
 
-        void storeInHash(unsigned int index, vector<float>* point, unsigned int pointID);
+        void storeInHash(unsigned int index, vector<float>* point, int pointID);
 
         pair<vector<float>,float> getVTParameters(unsigned int index);
 
-        priority_queue<pair<unsigned int, vector<float>*> > approximateKNN(unsigned int neighbours,
-            unsigned int index, vector<float>* point, unsigned int pointID);
+        priority_queue<pair<double, vector<float>*> > approximateKNN(unsigned int neighbours,
+            unsigned int index, vector<float>* point, int pointID);
 
-        priority_queue<pair<unsigned int, vector<float>*> > exactKNN(unsigned int neighbours, 
+        priority_queue<pair<double, vector<float>*> > exactKNN(unsigned int neighbours, 
             vector<float>* point);
             
-        set<vector<float>*> rangeSearch(unsigned int radius, unsigned int index, 
-            vector<float>* point, unsigned int pointID);
+        set<vector<float>*> rangeSearch(double radius, unsigned int index, 
+            vector<float>* point, int pointID);
 
         void deleteAllAllocatedPoints();
 };
